refactor(user_management): move user_roles data source into user package
The user_roles data source added in #454 lived in its own internal/user_roles/ package. Move it into internal/user/ alongside the other user resources so all user-related code shares one package.

Resolve package-level symbol collisions with the existing user package: drop the duplicate documentationSection constant and rename the data source's requiredScopes to dataSourceRequiredScopes. Update the provider registration to reference user.NewUserRolesDataSource.

No behavior change: the crowdstrike_user_roles data source type name and schema are unchanged.
