+++ internal/provider/provider.go
@@ -40,7 +40,6 @@ import (
 	"github.com/crowdstrike/terraform-provider-crowdstrike/internal/testconfig"
 	"github.com/crowdstrike/terraform-provider-crowdstrike/internal/user"
 	usergroup "github.com/crowdstrike/terraform-provider-crowdstrike/internal/user_group"
-	userroles "github.com/crowdstrike/terraform-provider-crowdstrike/internal/user_roles"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/function"
@@ -392,7 +391,7 @@ func (p *CrowdStrikeProvider) DataSources(ctx context.Context) []func() datasour
 		correlationrules.NewCorrelationRulesDataSource,
 		ngsiem.NewDataConnectorDataSource,
 		ngsiem.NewDataConnectorsDataSource,
-		userroles.NewUserRolesDataSource,
+		user.NewUserRolesDataSource,
 	}
 }
